+++ mobilitydb/sql/point/050_stbox.in.sql
@@ -102,24 +102,24 @@ CREATE FUNCTION stbox_t(tstzspan)
   AS 'MODULE_PATHNAME', 'Stbox_constructor_t'
   LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
 
-CREATE FUNCTION stbox(float8, float8, float8, float8, srid int DEFAULT 0)
+CREATE FUNCTION stbox(float, float, float, float, srid int DEFAULT 0)
   RETURNS stbox
   AS 'MODULE_PATHNAME', 'Stbox_constructor'
   LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
 
-CREATE FUNCTION stbox_z(float8, float8, float8, float8, float8, float8,
+CREATE FUNCTION stbox_z(float, float, float, float, float, float,
     srid int DEFAULT 0)
   RETURNS stbox
   AS 'MODULE_PATHNAME', 'Stbox_constructor_z'
   LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
 
-CREATE FUNCTION stbox_t(float8, float8, float8, float8, tstzspan,
+CREATE FUNCTION stbox_t(float, float, float, float, tstzspan,
     srid int DEFAULT 0)
   RETURNS stbox
   AS 'MODULE_PATHNAME', 'Stbox_constructor_t'
   LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
 
-CREATE FUNCTION stbox_zt(float8, float8, float8, float8, float8, float8,
+CREATE FUNCTION stbox_zt(float, float, float, float, float, float,
     tstzspan, srid int DEFAULT 0)
   RETURNS stbox
   AS 'MODULE_PATHNAME', 'Stbox_constructor_zt'
@@ -132,13 +132,13 @@ CREATE FUNCTION geodstbox_t(tstzspan)
   AS 'MODULE_PATHNAME', 'Geodstbox_constructor_t'
   LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
 
-CREATE FUNCTION geodstbox_z(float8, float8, float8, float8, float8, float8,
+CREATE FUNCTION geodstbox_z(float, float, float, float, float, float,
     srid int DEFAULT 4326)
   RETURNS stbox
   AS 'MODULE_PATHNAME', 'Geodstbox_constructor_z'
   LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
 
-CREATE FUNCTION geodstbox_zt(float8, float8, float8, float8, float8, float8,
+CREATE FUNCTION geodstbox_zt(float, float, float, float, float, float,
     tstzspan, srid int DEFAULT 4326)
   RETURNS stbox
   AS 'MODULE_PATHNAME', 'Geodstbox_constructor_zt'
